[Refactor] Extract ZipDownloader class
