+++ lib/soapy_bing/ads/bulk/campaigns.rb
@@ -3,8 +3,6 @@ module SoapyBing
   class Ads
     module Bulk
       class Campaigns
-        include Helpers::SSLVersion
-
         DEFAULT_ENTITIES = %w(CampaignTargets Ads).freeze
         POLLING_TRIES = 30
         NotCompleted = Class.new(StandardError)
@@ -18,7 +16,7 @@ module SoapyBing
         end
 
         def rows
-          @rows ||= Parsers::CSVParser.new(report_body).rows
+          @rows ||= Parsers::CSVParser.new(Helpers::ZipDownloader.new(result_file_url).read).rows
         end
 
         def result_file_url
@@ -42,20 +40,6 @@ module SoapyBing
             fetch_status
             raise NotCompleted if status['RequestStatus'] != 'Completed'
           end
-        end
-
-        def report_body
-          @report_body ||=
-            Zip::InputStream.open(download_io) do |archive_io|
-              file_io = archive_io.get_next_entry.get_input_stream
-              file_io.read
-            end
-        end
-
-        def download_io
-          https = URI.parse(result_file_url).scheme == 'https'
-          request_options = https ? { ssl_version: ssl_version } : {}
-          StringIO.new HTTParty.get(result_file_url, request_options).body
         end
 
         def download_request_id
