+++ lib/soapy_bing/helpers.rb
@@ -1,3 +1,3 @@
 # frozen_string_literal: true
 require 'soapy_bing/helpers/class_name'
-require 'soapy_bing/helpers/ssl_version'
+require 'soapy_bing/helpers/zip_downloader'
