--- spec/soapy_bing/helpers/ssl_version_spec.rb
@@ -1,26 +1,0 @@
-# frozen_string_literal: true
-RSpec.describe SoapyBing::Helpers::SSLVersion do
-  describe '#ssl_version' do
-    subject(:ssl_version) do
-      stub_const('MyClass', Class.new.include(described_class)).new.ssl_version
-    end
-
-    context 'when DEFAULT_SSL_VERSION is available' do
-      before do
-        stub_const('OpenSSL::SSL::SSLContext::METHODS', [described_class::DEFAULT_SSL_VERSION])
-      end
-
-      it 'responds with DEFAULT_SSL_VERSION' do
-        expect(ssl_version).to be MyClass::DEFAULT_SSL_VERSION
-      end
-    end
-
-    context 'when DEFAULT_SSL_VERSION is not available' do
-      before { stub_const('OpenSSL::SSL::SSLContext::METHODS', []) }
-
-      it 'responds with OpenSSL default version' do
-        expect(ssl_version).to be OpenSSL::SSL::SSLContext::DEFAULT_PARAMS[:ssl_version]
-      end
-    end
-  end
-end
